voronoi control 2D and 3D were seperated
