+++ examples/mechanical_square/mechanical_2D_poly_lin/cross_section_data.txt
@@ -1,0 +1,11 @@
+x, stress_x_fem, stress_x_fol, stress_y_fem, stress_y_fol, stress_xy_fem, stress_xy_fol
+0.0, 0.016417037695646286, 0.016433481127023697, 0.007035872433334589, 0.007042921148240566
+0.1111111111111111, 0.01840873993933201, 0.01851499266922474, 0.005487687885761261, 0.005558772012591362
+0.2222222222222222, 0.023959042504429817, 0.023864828050136566, 0.006139755714684725, 0.005954449065029621
+0.3333333333333333, 0.029465850442647934, 0.029324565082788467, 0.006814115215092897, 0.006653439253568649
+0.4444444444444444, 0.031889408826828, 0.03173741325736046, 0.007301948964595795, 0.007171220611780882
+0.5555555555555556, 0.02965548075735569, 0.02963581494987011, 0.008104627951979637, 0.00795120932161808
+0.6666666666666666, 0.01933516003191471, 0.01945171318948269, 0.004587389063090086, 0.004645001143217087
+0.7777777777777777, 0.018771853297948837, 0.01872217282652855, 0.004473820794373751, 0.0044075604528188705
+0.8888888888888888, 0.018999850377440453, 0.018950577825307846, 0.006071604322642088, 0.005990628153085709
+1.0, 0.01977470889687538, 0.0197382103651762, 0.00847487524151802, 0.008459233678877354
